// src/generated/prisma/models/Token.ts
export type TokenUncheckedCreateInput = {
  id?: string
  accessToken: string
  refreshToken: string
  revoked: boolean
  expired: boolean
  tokenType?: $Enums.TokenType
  userId: string
  createdAt?: Date | string
  updatedAt?: Date | string
}